fix(kotlin-sdk): invitation review hardening — durable slot allocator, app-scoped state, bearer routing
Implements the reviewed PR-feedback hardening for the invitations port
(the fixes for the three blocking #4284 review threads, which merged
without them):

- Durable, monotonic per-wallet identity-index allocation
  (IdentityIndexReservationStore + identity_index_state table seeded in
  the v9 migration; reserve-only-forward, exact-slot and
  asset-lock-resume semantics; RegistrationCoordinator owns allocation,
  controllers, invitation-operation permits, and the synchronous
  pendingRegistrationStarts gate so the network switch can never observe
  an idle window).
- App-scoped, deliberately non-saveable create/claim/contact/reclaim
  state machines with generation ids, SecretInvitationUri redaction, and
  recreation re-attach; network switch clears them; the Settings gate
  observes permits + in-flight states.
- Active-network manager guard before any key work at all four
  app-scope call sites.
- Fail-closed QR content classifier ahead of every SavedStateHandle
  write (custom + applink transports, any pk query key, percent-decoded,
  case-insensitive); claim-side FLAG_SECURE.
- u32 timestamps as Long through the persistence trampoline with the
  shared INVITATION_UPSERT_DESCRIPTOR constant (([B[BIJJJZI)I) used by
  both the call site and the descriptor verifier; InvitationEntity
  expiry/createdAt widened to Long (INTEGER affinity — no schema bump).
- Exact-row reclaim persistence (affected-rows == 1 checked on every
  marker/status write) + the canSubmitInvitationReclaim stale-snapshot
  gate.

JVM suites (475+ tests) + :app:assembleDebug green; JNI lib tests 38/38;
fmt clean. Instrumented tier (v8→v9 migration seeding + descriptor
resolution) still needs one emulator pass.

+++ packages/kotlin-sdk/sdk/src/main/kotlin/org/dashfoundation/dashsdk/ffi/NativePersistenceBridge.kt
@@ -489,7 +489,7 @@ abstract class NativePersistenceBridge {
     // ── Invitations (DIP-13) ──────────────────────────────────────────
 
     /**
-     * One `InvitationEntryFFI` upsert. Descriptor `([B[BIJIIZI)I`.
+     * One `InvitationEntryFFI` upsert. Descriptor `([B[BIJJJZI)I`.
      *
      * A non-zero return fails the persist round: `create_invitation` treats
      * an unrecorded invitation row as a hard error (a funded voucher with no
@@ -506,8 +506,8 @@ abstract class NativePersistenceBridge {
         outPoint: ByteArray,
         fundingIndex: Int,
         amountDuffs: Long,
-        expiryUnix: Int,
-        createdAtSecs: Int,
+        expiryUnix: Long,
+        createdAtSecs: Long,
         hasInviter: Boolean,
         status: Int,
     ): Int = 0
